requirements.txt
asgiref==3.4.1
Django==3.2.14
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-js-asset==2.0.0
djangorestframework==3.12.4
docutils==0.18
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
pkg_resources==0.0.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
yapf==0.32.0
